Add ConsoleObject::onCreation()
Interacting with Console is done by accessing ConsoleObjects by id - but
during execution of ConsoleObject::ConsoleObject(), the id does not yet
exist.

That means, exceptions are thrown if object wants to interact in
constructor.

So, when ConsoleObject-derived class wants to interact, it should be
done in onCreation(), or anywhere else. But not in constructor.